pyproject.toml
[project]
name = "food-recsys"
version = "0.1.0"
description = "A Food Recommendation System that suggests the best dishes based on user preferences and dietary restrictions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.0",
    "neo4j>=5.25.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "werkzeug>=3.0.0",
]

[project.scripts]
food-recsys = "main:main"
